refactor : CarInfo, Owner 데이터 Model 형태 수정
