+++ domain/src/main/java/com/gta/domain/model/UserProfile.kt
@@ -3,6 +3,6 @@ package com.gta.domain.model
 data class UserProfile(
     val id: String,
     val name: String,
-    val temp: Int,
+    val temp: Float,
     val image: String?
 )
